--- mineru/model/ocr/paddleocr2pytorch/__init__.py
@@ -1,1 +1,0 @@
-# Copyright (c) Opendatalab. All rights reserved.
